fix(storage): add voting_mechanism to write path, upgrade to f64 precision
The NewProposal struct and CreateProposalInputView were missing the
voting_mechanism and config fields (score_min, score_max, dots_per_voter,
quorum_percentage, passage_threshold), making it impossible to create
proposals with any mechanism other than the default 'consent'. Also
upgrades quorum_percentage and passage_threshold from f32/Float to
f64/Double for proper precision across all layers.

+++ elohim/elohim-storage/src/db/diesel_schema.rs
@@ -454,8 +454,8 @@ diesel::table! {
         score_min -> Nullable<Integer>,
         score_max -> Nullable<Integer>,
         dots_per_voter -> Nullable<Integer>,
-        quorum_percentage -> Nullable<Float>,
-        passage_threshold -> Nullable<Float>,
+        quorum_percentage -> Nullable<Double>,
+        passage_threshold -> Nullable<Double>,
     }
 }
